pyproject.toml
[project]
name = "fixer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ebooklib>=0.18",
    "epubcheck>=5.1.0",
    "rich>=13.9.4",
    "jpype1>=1.5.1",
]

[dependency-groups]
dev = [
    "pdbpp>=0.10.3",
    "pre-commit>=4.0.1",
]

[tool.ruff.lint]
select = ["I"]

[tool.uv.sources]
ebooklib = { git = "https://github.com/aerkalov/ebooklib" }
